src/requirements.txt
asyncpg==0.29.0
databases[asyncpg]==0.9.0
fastapi==0.115.6
google-api-python-client==2.160.0
google-auth-oauthlib==1.2.1
pydantic[email]==2.10.5
SQLAlchemy==2.0.36
requests==2.32.3